prepared terrain for libc

+++ kernel/include/timer.h
@@ -7,6 +7,6 @@
 
 void timer_init(uint32_t freq);
 void timer_disable(void);
-uint32_t timer_get(void);
+uint64_t timer_get(void);
 
 #endif // KERNEL_TIMER_H
